+++ src/Client.php
@@ -230,6 +230,8 @@ final class Client implements ClientInterface
             if (null !== $hint->stacktrace && null === $event->getStacktrace()) {
                 $event->setStacktrace($hint->stacktrace);
             }
+
+            $event->setExtra(array_merge($event->getExtra(), $hint->extra));
         }
 
         $this->addMissingStacktraceToEvent($event);